Variadic_Function
Variadic functions are functions that can take a variable number of arguments. In C programming, a variadic function adds flexibility to the program.
In this problem, you will implement three variadic functions named , sum() , max() and min() to calculate sums, minima, maxima of a variable number of arguments. The first argument passed to the variadic function is the count of the number of arguments, which is followed by the arguments themselves.